:recycle: [Refactor] 불필요한 파일 제거 루트 페이지에 날씨 관련 mvp 로 변경 #DEV
