--- src/components/textProvider/TextProvider.css
@@ -1,44 +1,0 @@
-.text-provider {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  gap: 1rem;
-  padding: 1.5rem;
-}
-
-.category-selector {
-  margin-bottom: 1rem;
-}
-
-.category-select {
-  padding: 0.5rem;
-  border-radius: 4px;
-  border: 1px solid #ddd;
-}
-
-.provider-button {
-  background-color: #0070f3;
-  color: white;
-  padding: 0.5rem 1rem;
-  border-radius: 4px;
-  border: none;
-  cursor: pointer;
-  transition: background-color 0.2s;
-}
-
-.provider-button:hover {
-  background-color: #0051b3;
-}
-
-.text-display {
-  background-color: #f5f5f5;
-  padding: 1rem;
-  border-radius: 4px;
-  margin-top: 1rem;
-}
-
-.provided-text {
-  color: #333;
-  font-size: 1rem;
-  margin: 0;
-}
